build: use pkgroll
