+++ package.json
@@ -22,11 +22,12 @@
 	"files": [
 		"dist"
 	],
-	"main": "dist/index.js",
+	"main": "./dist/index.js",
+	"types": "./dist/plugin-type.d.ts",
 	"scripts": {
 		"lint": "eslint .",
-		"build": "tsc",
-		"typecheck": "tsc --noEmit",
+		"build": "pkgroll",
+		"type-check": "tsc --noEmit",
 		"test": "pnpm build && pnpm test:wp4 && pnpm test:wp5",
 		"test:wp4": "pnpm --use-node-version=12.22.12 exec jest",
 		"test:wp5": "WEBPACK=5 pnpm --use-node-version=12.22.12 exec jest"
@@ -63,6 +64,7 @@
 		"jest": "^27.3.1",
 		"lint-staged": "^12.1.2",
 		"mini-css-extract-plugin": "^1.6.0",
+		"pkgroll": "^1.5.0",
 		"simple-git-hooks": "^2.8.0",
 		"tempy": "^1.0.1",
 		"typescript": "^4.8.3",
